proyek/requirements.txt
# git+https://github.com/facebookresearch/detectron2.git
torch==2.0.0
torchvision==0.15.1
opencv-python==4.6.0.66
matplotlib==3.5.3
gtts==2.3.2
playsound==1.2.2
mysql-connector-python==8.0.31